+++ test/amxo_include/test_include.h
@@ -67,5 +67,6 @@ void test_none_existing_include_file(void** state);
 void test_none_existing_optional_include_file(void** state);
 void test_recursive_include_detection(void** state);
 void test_include_absolute_path(void** state);
+void test_post_include(void** state);
 
 #endif // __TEST_INCLUDE_H__