+++ test/amxo_include/test_main.c
@@ -73,6 +73,7 @@ int main(void) {
         cmocka_unit_test(test_none_existing_optional_include_file),
         cmocka_unit_test(test_recursive_include_detection),
         cmocka_unit_test(test_include_absolute_path),
+        cmocka_unit_test(test_post_include),
     };
     return cmocka_run_group_tests(tests, NULL, NULL);
 }